Revert "Move stores into directory"
This reverts commit 4bc771da1edfefdeb1dc3c10d56a11b940d18331.
