+++ static/js/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var React = require('react');
-var VariableStore = require('./stores/variable-store');
+var VariableStore = require('./variable-store');
 var ModalRenderer = require('./modal-renderer');
 var Iframe = require('./iframe');
 var Sidebar = require('./sidebar');
