+++ static/js/variable-store.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var Store = require('../stores/store');
+var Store = require('./store');
 var $ = require('jquery');
 var _ = require('underscore');
